requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
lxml==4.9.3
numpy==1.26.2
outcome==1.3.0.post0
pandas==2.1.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
undetected-chromedriver==3.5.4
urllib3==2.1.0
websockets==12.0
wsproto==1.2.0
python-dotenv~=1.0.1
pygame~=2.6.1
webdriver-manager~=4.0.2